fix: Fix the button in Overview screen for adding a trip and repository to not call the collection on init
The floating action button now succesfully redirects to the AddTrip screen when pressed.

The repository now does not call db.collection().get() which was connecting to Firestore before authentification causing the Trips to not display after LogIn.
Remove test which was testing if the collection is called in the init method.
Adds user in the SignIn screen to remember user on succesful authentication.