Feat: Add navigation bar and screens for a trip
Previously we only displayed a list of the trips. This change allows
users to view some screens specific to a trip: Schedule, Activities and
Settings. They are accessible using a bottom bar. To return to the
Overview with all trips, I added a "Home" button on the top left of each
new screen. These screens currently display a prompt, but will be
implemented in future pull requests.